sql_runner/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-cloud-bigquery==3.25.0
pydantic==2.10.6